Iterates through root directory
This version should be able to go through a directory and get all of
the files out and create them in the current working directory.
However, it still needs to be able to go through the subdirectories
which i think should just be a call to goThroughDirectory() passing in
the start of the subdirectory and its size when a subdirectory is found
in the root directory.